requirements.txt
pyodbc==5.1.0
tabulate==0.9.0
openai==1.60.2
httpx==0.27.2
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-experimental==0.0.49
langchain-openai==0.0.5
pandas==2.2.2
SQLAlchemy==2.0.30
python-dotenv==1.0.1
streamlit
openpyxl      # Excel .xlsx
xlrd          # Excel .xls
odfpy         # OpenDocument .ods
pyyaml        # YAML files
pymysql       # MySQL
psycopg2-binary  # PostgreSQL
pyarrow       # Parquet + Feather
fastparquet   # Optional Parquet engine
